# balance-service/api/balance/balance_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v6.33.2
// source: api/balance/balance.proto

package balance

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationBalanceServiceGetBalanceHistory = "/api.balance.BalanceService/GetBalanceHistory"
const OperationBalanceServiceGetCurrentBalance = "/api.balance.BalanceService/GetCurrentBalance"
const OperationBalanceServiceRecordBalanceChange = "/api.balance.BalanceService/RecordBalanceChange"

type BalanceServiceHTTPServer interface {
	GetBalanceHistory(context.Context, *GetBalanceHistoryRequest) (*GetBalanceHistoryResponse, error)
	GetCurrentBalance(context.Context, *GetCurrentBalanceRequest) (*Balance, error)
	RecordBalanceChange(context.Context, *RecordBalanceChangeRequest) (*Balance, error)
}

func RegisterBalanceServiceHTTPServer(s *http.Server, srv BalanceServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/balance", _BalanceService_RecordBalanceChange0_HTTP_Handler(srv))
	r.GET("/account/{account_id}/balance", _BalanceService_GetCurrentBalance0_HTTP_Handler(srv))
	r.GET("/account/{account_id}/balance/history", _BalanceService_GetBalanceHistory0_HTTP_Handler(srv))
}

func _BalanceService_RecordBalanceChange0_HTTP_Handler(srv BalanceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RecordBalanceChangeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceServiceRecordBalanceChange)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RecordBalanceChange(ctx, req.(*RecordBalanceChangeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Balance)
		return ctx.Result(200, reply)
	}
}

func _BalanceService_GetCurrentBalance0_HTTP_Handler(srv BalanceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCurrentBalanceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceServiceGetCurrentBalance)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCurrentBalance(ctx, req.(*GetCurrentBalanceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Balance)
		return ctx.Result(200, reply)
	}
}

func _BalanceService_GetBalanceHistory0_HTTP_Handler(srv BalanceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetBalanceHistoryRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBalanceServiceGetBalanceHistory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetBalanceHistory(ctx, req.(*GetBalanceHistoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetBalanceHistoryResponse)
		return ctx.Result(200, reply)
	}
}

type BalanceServiceHTTPClient interface {
	GetBalanceHistory(ctx context.Context, req *GetBalanceHistoryRequest, opts ...http.CallOption) (rsp *GetBalanceHistoryResponse, err error)
	GetCurrentBalance(ctx context.Context, req *GetCurrentBalanceRequest, opts ...http.CallOption) (rsp *Balance, err error)
	RecordBalanceChange(ctx context.Context, req *RecordBalanceChangeRequest, opts ...http.CallOption) (rsp *Balance, err error)
}

type BalanceServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewBalanceServiceHTTPClient(client *http.Client) BalanceServiceHTTPClient {
	return &BalanceServiceHTTPClientImpl{client}
}

func (c *BalanceServiceHTTPClientImpl) GetBalanceHistory(ctx context.Context, in *GetBalanceHistoryRequest, opts ...http.CallOption) (*GetBalanceHistoryResponse, error) {
	var out GetBalanceHistoryResponse
	pattern := "/account/{account_id}/balance/history"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBalanceServiceGetBalanceHistory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BalanceServiceHTTPClientImpl) GetCurrentBalance(ctx context.Context, in *GetCurrentBalanceRequest, opts ...http.CallOption) (*Balance, error) {
	var out Balance
	pattern := "/account/{account_id}/balance"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBalanceServiceGetCurrentBalance))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *BalanceServiceHTTPClientImpl) RecordBalanceChange(ctx context.Context, in *RecordBalanceChangeRequest, opts ...http.CallOption) (*Balance, error) {
	var out Balance
	pattern := "/balance"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBalanceServiceRecordBalanceChange))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}