# balance-service/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: BalanceService API
    version: 0.0.1
paths:
    /account/{accountId}/balance:
        get:
            tags:
                - BalanceService
            operationId: BalanceService_GetCurrentBalance
            parameters:
                - name: accountId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Balance'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /account/{accountId}/balance/history:
        get:
            tags:
                - BalanceService
            operationId: BalanceService_GetBalanceHistory
            parameters:
                - name: accountId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: fromDate
                  in: query
                  description: Optional filters
                  schema:
                    type: string
                - name: toDate
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetBalanceHistoryResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /balance:
        post:
            tags:
                - BalanceService
            operationId: BalanceService_RecordBalanceChange
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RecordBalanceChangeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Balance'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        Balance:
            type: object
            properties:
                accountId:
                    type: string
                availableBalance:
                    type: string
                pendingBalance:
                    type: string
                currency:
                    type: string
                changeAmount:
                    type: string
                changeReason:
                    type: string
                balanceTimestamp:
                    type: string
                    format: date-time
                balanceDate:
                    type: string
                    format: date-time
        GetBalanceHistoryResponse:
            type: object
            properties:
                balances:
                    type: array
                    items:
                        $ref: '#/components/schemas/Balance'
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        RecordBalanceChangeRequest:
            type: object
            properties:
                accountId:
                    type: string
                availableBalance:
                    type: string
                pendingBalance:
                    type: string
                currency:
                    type: string
                changeAmount:
                    type: string
                changeReason:
                    type: string
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: BalanceService